==> 02-data.Rmd <==
# Data sources

Note that there is no downloading option for the data on NBA official website, so we need to scrape the data from the official website https://www.nba.com/stats/ using `GET()` from `httr` library.

Since the data source is the NBA's official website, and we scraped the data directly from it, it is considered reliable. 

In this project we want to focus on the data for the last regular season and playoffs, meaning the 2020-21 Regular Season and Playoffs. 

**2020-21 Regular Season Shots Data**

The scraping procedure for detailed shots data for 2020-21 Regular Season is as follows:

* For the detailed shooting data, we firstly land in the https://www.nba.com/stats/ webpage. Scrolling down, we can see a "Shotchart Search" section. Inputting one player's name, taking Stephen Curry during the 2021-22 season as an example, we can land on the page https://www.nba.com/stats/events/?flag=3&CFID=33&CFPARAMS=2021-22&PlayerID=201939&ContextMeasure=FGA&Season=2021-22&section=player&sct=hex. 

* In this page, we can see data with each row being a field goal that gives us information like the type of shot, whether it was successful, shot distance, etc. However, the exact location of the shot doesn't appear. 

* To access the detailed shots location data, we can right-click on the webpage and select "Inspect" and navigate to the "Network" tab. Then we can find the `shotchartdetail` data (if not appear, refresh page). Once find it, right-click it and open it in a new tab, so we can get the URL.

* After getting the URL, we can use the url to import the data on the webpage into our R studio using `GET()` function.

```{r}
library(tidyverse)
library(httr)
```

```{r}

headers = c(
  `Connection` = 'keep-alive',
  `Accept` = 'application/json, text/plain, */*',
  `x-nba-stats-token` = 'true',
  `X-NewRelic-ID` = 'VQECWF5UChAHUlNTBwgBVw==',
  `User-Agent` = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/78.0.3904.87 Safari/537.36',
  `x-nba-stats-origin` = 'stats',
  `Sec-Fetch-Site` = 'same-origin',
  `Sec-Fetch-Mode` = 'cors',
  `Referer` = 'https://stats.nba.com/players/leaguedashplayerbiostats/',
  `Accept-Encoding` = 'gzip, deflate, br',
  `Accept-Language` = 'en-US,en;q=0.9'
)

# define a function to grab shots data
fetch_shots = function(player_id, season, season_type) {
  
  # GET request
  request = GET(
    "http://stats.nba.com/stats/shotchartdetail",
    query = list(
      PlayerID = player_id, # specific id OR 0 for all players
      Season = season, # game season in format such as "2020-21"
      SeasonType = season_type, #"Regular Season" OR "PlayOff"
      PlayerPosition = "",
      ContextMeasure = "FGA",
      DateFrom = "",
      DateTo = "",
      GameID = "",
      GameSegment = "",
      LastNGames = 0,
      LeagueID = "00",
      Location = "",
      Month = 0,
      OpponentTeamID = 0,
      Outcome = "",
      Period = 0,
      Position = "",
      RookieYear = "",
      SeasonSegment = "",
      TeamID = 0,
      VsConference = "",
      VsDivision = ""
    ),
    add_headers(headers)
  )
  
  stop_for_status(request)
  
  # grab requested data
  data = content(request)
  
  # get the desired data
  raw_shots_data = data$resultSets[[1]]$rowSet
  col_names = tolower(as.character(data$resultSets[[1]]$headers))
  
  # preprocess raw data
  if (length(raw_shots_data) == 0) {
    shots = data.frame(
      matrix(nrow = 0, ncol = length(col_names))
    )
  } else {
    shots = data.frame(
      matrix(
        unlist(raw_shots_data),
        ncol = length(col_names),
        byrow = TRUE
      )
    )
  }
  
  shots = as_tibble(shots)
  names(shots) = col_names
  
  return(shots)
}

shotchartdetail_regular = fetch_shots(0,"2020-21","Regular Season")
```

```{r}
write.csv(shotchartdetail_regular, "shotchartdetail_regular.csv",row.names = FALSE)
```

After building the function to grab shots data using `GET()` and doing simple data preprocessing, our final 2020-21 Regular Season Shots Data contains in total 24 columns and around 200,000 observations. 

The variables include player information (`playerID`, `player_name`), team information (`tean_ID`, `team_name`), game information (`game_ID`, `game_date`), and shot information (`period`, `minutes remaining` and `seconds remaining`, whether shot successful - `event_type`, shot XY location, `shot_distance`, `shot_zone`, `shot_area`, etc.). All variables are in either string or integer format.

**2020-21 Playoffs Shots Data**

Using a similar way, we can get the detailed shots data for Playoffs. 
```{r}
shotchartdetail_playoffs = fetch_shots(0,"2020-21","Playoffs")
write.csv(shotchartdetail_playoffs, "shotchartdetail_playoffs.csv",row.names = FALSE)
```

Our 2020-21 Playoffs Shots Data contain a total of 14771 observations, and the same variables as the 2020-21 Regular Season Shots Data.
