==> 04-missing.Rmd <==
# Missing Values

```{r include=FALSE}
library(pander)
library(knitr)
library(tidyverse)
library(jsonlite)
library(httr)
library(scales)
library(patchwork)
library(readr)
```

We will look at the missing pattern of detailed shots data for NBA 2020-21 Regular Season and Playoffs. 

**2020-21 Regular Season Shots Data**

After getting the data out in the data sources portion, here we firstly have a look at the head of the NBA 2020-21 Regular Season data: 

```{r}
shot_data = read_csv("shotchartdetail_regular.csv")
shot_data_playoff = read_csv("shotchartdetail_playoffs.csv")
head(shot_data,3)  %>% pander()
```

It seems that there is actually no missing values in our first 5 rows of data. To examine the whole dataset, we firstly count the total number of missing value:
```{r}
as_tibble(sum(is.na(shot_data))) %>% pander()
```

To further confirm our hypothesis that our dataset contains no missing data, we then draw a missing value plot:

```{r regular_missing}
# Input: df: dataframe to plot missing value; 
#        percent: indicator whether to output percentage (FALSE: count, TRUE: percent)
missing_plots <- function(df, percent = TRUE){
  
  # determine the level of variable based on # of missing rows
  var_levels <- data.frame(count = colSums(is.na(df))) %>%
    rownames_to_column("variable") %>%
    arrange(desc(count),variable) %>%
    pull(variable)
  
  # get the missing patterns
  missing_patterns <- data.frame(is.na(df)) %>%
  group_by_all() %>%
  count(name = "count", sort = TRUE) %>%
  ungroup()
  
  main <- missing_patterns %>%
    select(-count) %>%
    rowid_to_column("index") %>%
    rowwise() %>% 
    mutate(complete_case = sum(c_across(-index))) %>%
    mutate(complete_case = ifelse(complete_case == 0,TRUE,FALSE)) %>% # find complete cases
    gather("variable","missing_pattern",- c(index, complete_case)) %>% # make data in long form
    ggplot(aes(x = factor(variable, levels = var_levels), 
               y=reorder(index,-index), 
               fill = missing_pattern, 
               alpha = complete_case)) +
    geom_tile(color = "white", lwd = 0.5, linetype = 1)+
    geom_text(aes(label=
                    ifelse(complete_case==TRUE 
                             & variable == var_levels[as.integer(length(colnames(df))/2)+1],
                             "Complete Cases", "")))+ # label "Complete Cases" in middle
    scale_fill_manual(values=c("grey","purple4"))+ # define fill color 
    scale_alpha_manual(values=c(0.5,1))+ # define alpha
    ylab("missing cases") + xlab("variables") + # define x,y title
    theme_classic()+ 
    theme(legend.position = "none") # remove legend
  
  
  upper_c <- data.frame(count = colSums(is.na(df))) %>%
    rownames_to_column("variable") %>%
    ggplot(aes(x = factor(variable, levels = var_levels), y = count))+
    geom_col(fill = "steelblue", alpha = 0.5)+
    ylab(ifelse(percent == TRUE, "% rows missing", "# rows missing")) +
    theme_bw()+
    theme(panel.grid.major.x = element_blank(), # only keep horizontal lines
          axis.title.x=element_blank()) # remove x-axis title
  
  upper_p <- data.frame(count = colSums(is.na(df))) %>%
    mutate(count = count/nrow(df)*100) %>%
    rownames_to_column("variable") %>%
    ggplot(aes(x = factor(variable, levels = var_levels), y = count))+
    geom_col(fill = "steelblue", alpha = 0.5)+
    ylab(ifelse(percent == TRUE, "% rows missing", "# rows missing")) +
    theme_bw()+
    theme(panel.grid.major.x = element_blank(), # only keep horizontal lines
          axis.title.x=element_blank()) # remove x-axis title
  
  left <- missing_patterns %>%
  rowid_to_column("index") %>%
  rowwise() %>% 
  mutate(complete_case = sum(c_across(-c(index,count)))) %>%
  mutate(complete_case = ifelse(complete_case == 0,TRUE,FALSE)) %>%
  select(index,count,complete_case) %>%
  mutate(count = ifelse(percent == TRUE, count/nrow(df)*100, count)) %>%
  ggplot(aes(x = reorder(index,-index), y = count, alpha = complete_case))+
  geom_col(fill = "steelblue")+
  scale_alpha_manual(values=c(0.5,1))+
  ylab(ifelse(percent == TRUE, "% rows", "# rows")) +
  theme_bw()+
  coord_flip()+
  theme(panel.grid.major.y = element_blank(), # only keep horizontal lines
        axis.title.y=element_blank(), # remove x-axis title
        legend.position = "none") 
  
  if (percent == TRUE) {
    return (upper_p+main+left+ plot_layout(design = c( area(1, 1, 1, 3),area(2, 1, 4, 3), area(2, 4, 4, 4)))+plot_annotation(title = "Missing Value Plot with Percent of Missing Value"))
  }else{
    return (upper_c+main+left+ plot_layout(design = c( area(1, 1, 1, 3),area(2, 1, 4, 3), area(2, 4, 4, 4)))+plot_annotation(title = "Missing Value Plot with Count of Missing Value"))
  }
}

missing_plots(shot_data, FALSE)
```

From the plot, we can see that all rows are "completed cases", and the total number of missing values for each variable is 0, suggesting that there is no missing data.

**2020-21 Playoffs Shots Data**

Since there is no missing data in regular season, we would expect that there is also no missing data in playoffs. Firstly, we check the total number of missing data:
```{r}
as_tibble(sum(is.na(shot_data_playoff))) %>% pander()
```

Now we plot the missing value plot:
```{r playoff_missing}
missing_plots(shot_data_playoff, FALSE)
```

From the plot, similar as before, we can see that all rows are "completed cases", and the total number of missing values for each variable is 0.

Therefore, as expected, our results indicates that both of our dataset for regular season and playoffs contain no missing data. Therefore, we can finally conclude that there is no missing data in our datasets and proceed to next steps.
