==> 05-results.Rmd <==
# Results

```{r include=FALSE}
library(tidyverse)
library(ggplot2)
library(GGally)
library(vcd)
library(RColorBrewer)
library(ggalluvial)
#devtools::install_github("timelyportfolio/parcoords")
library(parcoords)
library(pander)
library(ggridges)
library(patchwork)
library(prismatic)
library(extrafont)
library(cowplot)
library(dplyr)
library(hexbin)
library(jsonlite)
library(httr)
library(scales)
library(RCurl)
library(jpeg)
library(ggalluvial)

```

```{r}
shot_data = read_csv("shotchartdetail_regular_updated.csv")
shot_data_playoff = read_csv("shotchartdetail_playoffs.csv")
```


In this chapter, we will proceed to come up with the answers and our conclusions to each of the questions we proposed in the introduction section. Here are our questions proposed in the introduction chapter:

1. (On League Level)
We will give an overall glance at the dataset for the entire league of last season to have a basic understanding of the shooting behaviors.

2. (On Team Level)
Basketball is a team-played game and each team has its own style and culture. From the perspective of the team, we are going to look for some differences between shooting behaviors preferences and learn the relationship between the shooting patterns and period, field goal percentage, etc.

3. (On Player Level)
Sometimes a single player could make a huge impact on games. We choose the three most popular players, Stephen Curry, Lebron James, and Giannis Antetokounmpo to investigate their shooting choices and look at how other variables could influence their shooting decision such as in different periods of the game.

## League Shooting Pattern Analysis

### Distribution of Shot Distance

```{r shoot_dist}
density_graph <- shot_data %>%
  ggplot(aes(x = shot_distance)) + 
  geom_histogram(aes(y = ..density..), bins = 25)+
  geom_density(aes(color = "density curve"))+ 
  labs(title = "Distribution of Shot Distance", x = "Shot Distance (in)", y = "Density")


density_ridges_zone <- shot_data %>% 
  ggplot(aes(x = shot_distance, y=shot_zone_basic)) +
  geom_density_ridges(alpha = .5) + 
  labs(title = "Ridgeline Plots of Shot Distance in Different Zones", x = "Shot Distance (in)", y = "Zone")

shot_dist_bar <- shot_data %>%
  ggplot(aes(x= shot_distance, y = event_type)) + 
  geom_boxplot()+
  labs(title = "Boxplots of Shot Result vs. Shot Distance", x = "Shot Distance (in)", y = "Shot Result")

density_graph + density_ridges_zone +shot_dist_bar+plot_layout(ncol=1)
```

From the first density histogram, we can see that the distribution of shot distance has two apparent peaks. Most shots are made around a distance of 0 inches and 25 inches. According to the Ridgeline Plots, we can conclude that the shots made around 0-inch distance are mainly in the restricted area, and the shots made around 25-inch distance are all kinds of 3-point shots. This graph shows that modern basketball tends to shoot less mid-range - either shooting close to the basket or making more 3-point shots - to try to maximize points per play.

Taking a look at the last boxplot with the distribution of shot distance according to shot result, it is obvious that missed shots have more outliers with larger distances. This situation makes sense because it is harder to make a long-distanced shot compared to a short-distanced one.

### Relationship Between Shot Result and Shot Distance

Since in the previous part we realized that missed shots have more outliers with larger distance, we want to explore whether the shot distance will affect whether the shot is successfully made.

```{r mosaic}
shot_dis_mos <- shot_data %>%
  select(c("event_type","shot_zone_range")) %>%
  mutate(shot_zone_range = factor(shot_zone_range, levels = c("Less Than 8 ft.", "8-16 ft.", "16-24 ft.","24+ ft.","Back Court Shot"))) %>%
  group_by(event_type,shot_zone_range) %>%
  summarize(Freq=n(),.groups = 'drop')

mosaic(event_type~ shot_zone_range, direction = c("v","h"), shot_dis_mos,
       highlighting_fill = c("cornflowerblue","grey80"),
       labeling = labeling_border(tl_labels = c(FALSE, TRUE), rot_labels = c(30,0)),
       main = "Mosaic Plot on Shot Distance and Shot Result")
```

From the mosaic plot, we can see that as the shot distance increase, fewer shots are successfully made and more shots are missed. To further confirm our hypothesis that shot distance will affect the shot result, we perform chi-square tests to test for associations between the shot distance and shot result.

Null Hypothesis: `Shot Distance` and `Shot Result` are independent.

Alternative Hypothesis: `Shot Distance` and `Shot Result` are NOT independent.

```{r}
shot_dis_mos_wide <- shot_dis_mos %>%
  pivot_wider(names_from = event_type, values_from = Freq, values_fill = 0)

shot_dis_mos_mtx <- as.matrix(shot_dis_mos_wide[2:3])

rownames(shot_dis_mos_mtx) = shot_dis_mos_wide$shot_zone_range

pander(chisq.test(shot_dis_mos_mtx, correct = FALSE))
```

Since the p-value is less than 0.01, meaning the probability of us accepting the null hypothesis is less than 1%, so we can reject the null hypothesis that `Shot Distance` and `Shot Result` are independent. Therefore, it means that there is a relation between `Shot Distance` and `Shot Result`.

### Shot Attempted in Different Periods

Every NBA game has four regular periods to decide which team is the winner (there would be overtime periods if the score is tied), so here we will ignore those overtime period and see generally how well all players shoot in different periods.

```{r period_bar}
shot_cnt <- shot_data %>%
  select(c("event_type","shot_type","period")) %>%
  filter(period <= 4) %>%
  mutate(period = factor(period)) %>%
  group_by(event_type,shot_type,period) %>%
  summarize(count=n(),.groups = 'drop') %>%
  ggplot(aes(x = period, y = count, fill = event_type))+
  geom_col() +
  facet_wrap(~shot_type)+
  labs(title = "Stacked Bar Plot of Shot Count in Different Period")

shot_freq <- shot_data %>%
  select(c("event_type","shot_type","period")) %>%
  filter(period <= 4) %>%
  mutate(period = factor(period)) %>%
  group_by(period, shot_type,event_type) %>%
  summarize(count=n(),.groups = 'drop') %>%
  group_by(period,shot_type) %>%
  mutate(freq = count/sum(count)) %>%
  ungroup() %>%
  ggplot(aes(x = period, y = freq, fill = event_type))+
  geom_col() +
  facet_wrap(~shot_type)+
  labs(title = "Stacked Bar Plot of Shot Frequency in Different Period")
shot_cnt/shot_freq
```

As shown in the second bar plots above, for both 2-point and 3-point shots, the field goal percentage stays similar for each period because the colored area of `Made Shot` and `Missed Shot` stays similar for all periods. For 2-point shots, the shot result approximately accounts for 50% of the area respectively, but for the 3-point field goals, `Made Shot` is less than it for 2-point shots, which make sense because making 3-point shots is more difficult than making 2-point shots as we discussed in the previous section that distance affects the shooting result. Also, we can see that for 2-point shots, the total shots decrease as the period increases. This is a normal phenomenon to us because players would be too exhausted to shoot near the hoop when the game approaches the end. Generally speaking, this overall situation for the entire league performance is as expected.

## Team Shooting Pattern Analysis

At next step, we would go deeper onto the shooting performance on team level.

### Total Shots for Each Team

Although in the previous section, we realized that there is no striking difference in total number and FG% for 2-point and 3-point shots in each period for the entire league, we somewhat expect this shooting pattern to differ team by team. Firstly, to get a general understanding of how many 2-point and 3-point shots each team attempted for season 2020-21, we create a Cleveland Dot Plot as follows:

```{r team_pref}
shot_data %>%
  select(team_name, shot_type) %>%
  group_by(team_name,shot_type) %>%
  summarize(n = n(),.groups = 'drop') %>%
  pivot_wider(values_from = n, names_from = shot_type) %>%
  rename_all(make.names) %>%
  rename(two_point_shots = X2PT.Field.Goal, three_point_shots = X3PT.Field.Goal) %>%
  mutate(shots_sum = two_point_shots+ three_point_shots) %>% # calculate total
  arrange(desc(shots_sum)) %>% # order by total
  pivot_longer(cols = c("two_point_shots", "three_point_shots","shots_sum")) %>% # tidy the data
  ggplot(aes(x = value, y = reorder(team_name, value), color = name))+
  geom_point()+
  theme_linedraw()+ 
  labs(title = "2-point and 3-point Total Shots for Each Team", x = "count", y = "Team Name")
```

From the Cleveland Dot Plot, we can see that the total shots made by each team for the entire season don't vary a lot, ranging from around 6000 to 6500. However, different teams have different preferences for 2-point shots and 3-point shots. For example, Utah Jazz has almost the same counts for 2-point shots and 3-point shots with the most 3-point shots in the league. On the contrary, San Antonio Spurs has the least count of 3-point shots and the most count of 2-point shots in the league. Therefore, we can conclude that teams vary significantly in their choice for making 2-point or 3-point shots.

### Three-point Preference by Period

After generally seeing the preference of different teams, now we can explore the movement of preference for 2-point shots and 3-point shots period by period. To see the preference change, here we calculate the percentage of 3-point shots among all shots made in each period for each team ($\text{percentage} = \frac{\text{3-point shots attempted}}{\text{all shots attempted}}$). If the percentage increase, it means that the team prefers to make more 3-point shots.

```{r alluvial_pref_period}
getPalette = colorRampPalette(brewer.pal(9, "Set1"))

period_shots <- shot_data %>%
  filter(period <= 4) %>%
  select(team_name, shot_type, period) %>%
  mutate(period = factor(period)) %>%
  group_by(team_name, shot_type, period) %>%
  summarize(n = n(),.groups = 'drop') %>%
  pivot_wider(names_from = shot_type, values_from = n) %>%
  rename_all(make.names) %>%
  mutate(p = X3PT.Field.Goal/(X3PT.Field.Goal + X2PT.Field.Goal)) %>%
  mutate(three_point_percentage_cut = cut(p, breaks=c(-Inf,0.3,0.35,0.4,0.45,0.5,Inf), 
                            labels=c("<30%", "30%-35%", "35%-40%", "40%-45%","45%-50%","50%+"))) %>%
  mutate(three_point_percentage_cut = as.character(three_point_percentage_cut)) %>% 
  select(-c(X2PT.Field.Goal,X3PT.Field.Goal,p)) %>%
  pivot_wider(names_from = period, values_from = three_point_percentage_cut) 
  
colnames(period_shots) <- c("Name", "Period1", "Period2", "Period3","Period4")

period_shots%>%
  ggplot(aes( axis1 = Period1, axis2 = Period2, axis3 = Period3, axis4 = Period4))+
  geom_alluvium(aes(fill = Name), width = 1/12)+
  geom_stratum(width = 1/12, fill = "grey80", color = "dark grey")+
  geom_label(stat="stratum",
             aes(label = after_stat(stratum)))+
  ggtitle("Alluvial Diagram of 3-point Percentage Change")+
  theme_void()+
  scale_fill_manual(values = getPalette(30))+
  theme(legend.position="bottom",legend.title = element_text(size=4), legend.text = element_text(size=4))

```

*(Here Each color represents one team. The percentage represent the percentage of 3-point shots of the total number of shots attempted each period. The first axis means 1st period, and so on.)*

From the graph, we can see that, indeed, most of the teams will have different preferences for 2-point and 3-point shots regarding different periods of a game. For example, Brooklyn Nets experienced the largest change in the percentage of 3-point shots among other teams: at 1st period, it has a percentage between 40% and 35%, and during the 2nd period, the percentage decreased to 35% - 40%. Then during the 3rd period, the percentage increased to 40%-45% and finally reached 45%-50% at the 4th period.

A general trend is that, as the game approaches 4th period, teams tend to make more 3-point shots compared to their performance in 1st period.

### Field Goal Percentage

After getting a general understanding of types and shots attempted for each team, now we want to focus on the Field Goal Percentage ($\frac{\text{ total shots made}}{\text{total shots attempted}}$).

Here we will use a diverging bar chart to compare the Field Goal % for 2-point and 3-point shots for each team, ordering by Field Goal % for 3-point shots:
```{r team_fgp}
FG_shot_type <- shot_data %>%
  select(team_name, shot_type,shot_attempted_flag,shot_made_flag) %>%
  group_by(team_name,shot_type) %>%
  summarize(attempts = sum(shot_attempted_flag),
            goals = sum(shot_made_flag), .groups = 'drop') %>%
  mutate(FGp = (goals/attempts)*100) %>%
  select(team_name,shot_type,FGp) %>%
  pivot_wider(values_from = FGp, names_from = shot_type)

HH::likert(team_name~., FG_shot_type, positive.order = TRUE,
           main = "FG% for 2-point and 3-point Shots for Each Team", 
           xlab = "Percent", ylab = "Team Name")
```

From the plot, we can see that, generally, 3-point Field Goal is less than 2-point Field Goal, which makes sense because 3-point shots are harder to score regarding the distance.

It is difficult to see if the FG% for 2-point or 3-point corresponds to the team's preference. Based on our previous conclusion, Utah Jazz has the most number of 3-point shots attempted, and San Antonio Spurs were concluded as the team least preferring 3-point shots. However, the FG% for 3-point shots are not the highest for Utah Jazz and not the lowest for San Antonio Spurs. This situation may be affected by the fact that, since the team prefers 3-point shots, they make more, and they miss more, so the FG% will not be the highest. Therefore, teams with more than three attempts can try to put more effort into the three-point practice so that the scores could be maximized.

## Individual Player Analysis

After analyzing the performance of teams, we would like to get deeper into an individual player. Since there are over 300 players in the league, we are not able to analyze each player one by one, so it is necessary to focus on the most brilliant stars.

### Shooting Behavior of Stephen Curry

Since in Part 5.1.1 we discovered that the league now generally prefers making either 2-point shots near hoop or 3-point shots with fewer mid-range shots, here we firstly explore the shooting behavior of Stephen Curry. As a ground-breaking player, Stephen Curry showed the world how explosive his scoring ability was and he is making an NBA record of three-pointers made while this project is getting done. Analyzing his shooting behavior would be a great angle to start.

```{r curry}
library(RCurl)
shot_steph <- shot_data %>%
  filter(player_name == "Stephen Curry")
# import court background image
courtImg.URL <- "https://thedatagame.files.wordpress.com/2016/03/nba_court.jpg"
court <- rasterGrob(readJPEG(getURLContent(courtImg.URL)),
           width=unit(1,"npc"), height=unit(1,"npc"))
curry_heat <- ggplot(shot_steph, aes(x=loc_x, y=loc_y)) + 
  annotation_custom(court, -250, 250, -50, 420) +
  geom_hex(bins = 40) +
  scale_fill_viridis_c(option = "inferno",direction = -1) +
  xlim(-250, 250) +
  ylim(-50, 420)+
  labs(title = "Curry's Shot Count")+
  theme_void()+
  theme(legend.position="bottom")
# League average FG% on different zones
league_avg <- shot_data %>%
  select(shot_zone_basic, shot_attempted_flag,shot_made_flag) %>%
  group_by(shot_zone_basic) %>%
  summarize(shot_attempted = sum(shot_attempted_flag), shot_made = sum(shot_made_flag)) %>%
  mutate(FGp = shot_made/shot_attempted) %>%
  select('FGp')
# Curry's FG% on different zones
shot_steph_fgp <- shot_steph %>%
  select(shot_zone_basic, shot_attempted_flag,shot_made_flag) %>%
  group_by(shot_zone_basic) %>%
  summarize(shot_attempted = sum(shot_attempted_flag), shot_made = sum(shot_made_flag)) %>%
  mutate(FGp = shot_made/shot_attempted)
# Compare Curry's FG% on different zones with League Average
shot_steph_fgp$league_avg <- league_avg$FGp
shot_steph_fgp$FGp_diff = (shot_steph_fgp$FGp - shot_steph_fgp$league_avg)*100
curry_fg <- shot_steph %>%
  mutate(FG_comp = case_when(shot_zone_basic == "Above the Break 3" ~ shot_steph_fgp$FGp_diff[1],
                             shot_zone_basic == "Backcourt" ~ shot_steph_fgp$FGp_diff[2], 
                             shot_zone_basic == "In The Paint (Non-RA)" ~ shot_steph_fgp$FGp_diff[3], 
                             shot_zone_basic == "Left Corner 3" ~ shot_steph_fgp$FGp_diff[4], 
                             shot_zone_basic == "Mid-Range" ~ shot_steph_fgp$FGp_diff[5], 
                             shot_zone_basic == "Restricted Area" ~ shot_steph_fgp$FGp_diff[6],
                             shot_zone_basic == "Right Corner 3" ~ shot_steph_fgp$FGp_diff[7])) %>%
  ggplot(aes(x=loc_x, y=loc_y)) + 
  annotation_custom(court, -250, 250, -50, 420) +
  geom_point(aes(color = FG_comp)) +
  scale_color_gradient2(midpoint = 0, low = "cornflowerblue", mid = "cornsilk2",
                            high = "brown3", name = "FG% compared to league average" )+
  xlim(-250, 250) +
  ylim(-50, 420)+
  labs(title = "Curry's FG% to League Average")+
  theme_void()+
  theme(legend.position="bottom")

curry_heat | curry_fg
```

These two plots use a graph of a basketball court to visualize the performance of Stephen Curry at each point on the court and it divides the court into multiple zones. The color in the left graph indicates the total number of shots made by Stephen Curry, the color in the right graph indicates the field goal percentage of Stephen Curry in that specific zone compared to the average level of the league. It is red if the player shoots above the league average and blue if lower than the league average.

In the plot above, we could first notice that Stephen Curry tries most of his shots in the paint (a rectangle area nearest to the basket) and out of a three-point line. Again, this shooting behavior is a classical example of modern basketball: try less on the mid-range shot and more 2-point shots near the hoop and 3-point shots. Another characteristic of his shooting is the range of his threes. We could even see multiple points scattered a long distance from the three-point line, which makes it harder to guard him.

On the other hand, we could still see that Stephen has his weakness, which is the left corner. His left-corner threes are deep blue so it would be a great strategy to limit him to shoot at that place. However, we think he can develop because he shoots very well on the other corner.


### Shooting Pattern in Playoffs of LeBron James

We could always say that games of the regular season do not matter because being a championship depends on the performance in playoffs. So many top players choose to take as much rest as they can in the regular season so that they could save more energy for the playoffs. Therefore, it would be great if we can find some differences in shooting behavior in the regular season and playoffs. As a die-hard Lakers fan, I would choose to analyze LeBron James.

```{r lebron_shot}
shot_lebron <-  shot_data %>%
  filter(player_name == "LeBron James")

shot_lebron_playoff <-  shot_data_playoff %>%
  filter(player_name == "LeBron James")

# League playoff average FG% on different zones
league_avg_playoff <- shot_data_playoff %>%
  select(shot_zone_basic, shot_attempted_flag,shot_made_flag) %>%
  group_by(shot_zone_basic) %>%
  summarize(shot_attempted = sum(shot_attempted_flag), shot_made = sum(shot_made_flag)) %>%
  mutate(FGp = shot_made/shot_attempted) %>%
  select('FGp')

# LeBron's FG% on different zones
shot_lebron_fgp <- shot_lebron %>%
  select(shot_zone_basic, shot_attempted_flag,shot_made_flag) %>%
  group_by(shot_zone_basic) %>%
  summarize(shot_attempted = sum(shot_attempted_flag), shot_made = sum(shot_made_flag)) %>%
  mutate(FGp = shot_made/shot_attempted)

# LeBron's FG% on different zones playoff
shot_lebron_fgp_playoff <- shot_lebron_playoff %>%
  select(shot_zone_basic, shot_attempted_flag,shot_made_flag) %>%
  group_by(shot_zone_basic) %>%
  summarize(shot_attempted = sum(shot_attempted_flag), shot_made = sum(shot_made_flag)) %>%
  mutate(FGp = shot_made/shot_attempted)

# Compare LeBron's FG% on different zones with League Average
shot_lebron_fgp$league_avg <- league_avg$FGp[-2]
shot_lebron_fgp$FGp_diff = (shot_lebron_fgp$FGp - shot_lebron_fgp$league_avg)*100

# Compare LeBron's Palyoff FG% on different zones with League Average
shot_lebron_fgp_playoff$league_avg <- league_avg_playoff$FGp[-2]
shot_lebron_fgp_playoff$FGp_diff = (shot_lebron_fgp_playoff$FGp - shot_lebron_fgp_playoff$league_avg)*100

lebron_fg <- shot_lebron %>%
  mutate(FG_comp = case_when(shot_zone_basic == "Above the Break 3" ~ shot_lebron_fgp$FGp_diff[1],
                             shot_zone_basic == "In The Paint (Non-RA)" ~ shot_lebron_fgp$FGp_diff[2], 
                             shot_zone_basic == "Left Corner 3" ~ shot_lebron_fgp$FGp_diff[3], 
                             shot_zone_basic == "Mid-Range" ~ shot_lebron_fgp$FGp_diff[4], 
                             shot_zone_basic == "Restricted Area" ~ shot_lebron_fgp$FGp_diff[5],
                             shot_zone_basic == "Right Corner 3" ~ shot_lebron_fgp$FGp_diff[6])) %>%
  ggplot(aes(x=loc_x, y=loc_y)) + 
  annotation_custom(court, -250, 250, -50, 420) +
  geom_point(aes(color = FG_comp)) +
  scale_color_gradient2(midpoint = 0, low = "cornflowerblue", mid = "cornsilk2",
                            high = "brown3", name = "FG% compared to league average" )+
  xlim(-250, 250) +
  ylim(-50, 420)+
  labs(title = "Regular Season: LeBron's FG% to Average")+
  theme_void()+
  theme(legend.position="bottom")

lebron_fg_playoff <- shot_lebron_playoff %>%
  mutate(FG_comp = case_when(shot_zone_basic == "Above the Break 3" ~ shot_lebron_fgp_playoff$FGp_diff[1],
                             shot_zone_basic == "In The Paint (Non-RA)" ~ shot_lebron_fgp_playoff$FGp_diff[2], 
                             shot_zone_basic == "Left Corner 3" ~ shot_lebron_fgp_playoff$FGp_diff[3], 
                             shot_zone_basic == "Mid-Range" ~ shot_lebron_fgp_playoff$FGp_diff[4], 
                             shot_zone_basic == "Restricted Area" ~ shot_lebron_fgp_playoff$FGp_diff[5],
                             shot_zone_basic == "Right Corner 3" ~ shot_lebron_fgp_playoff$FGp_diff[6])) %>%
  ggplot(aes(x=loc_x, y=loc_y)) + 
  annotation_custom(court, -250, 250, -50, 420) +
  geom_point(aes(color = FG_comp)) +
  scale_color_gradient2(midpoint = 0, low = "cornflowerblue", mid = "cornsilk2",
                            high = "brown3", name = "FG% compared to league average" )+
  xlim(-250, 250) +
  ylim(-50, 420)+
  labs(title = "Playoffs: LeBron's FG% to Average")+
  theme_void()+
  theme(legend.position="bottom")

lebron_fg|lebron_fg_playoff + plot_annotation(title = "LeBron FG% Compared to League Average")
```

By looking at the regular season shot chart on the left, it is obvious that LeBron did fantastic in the paint, which shows that his advantage near the basketball still exists. On contrary, he also tried a lot of mid-range shots and threes but the results were not ideal, except that his shooting percentage on the left corner is extremely high, where Stephen Curry was not good at.

Comparing the performance in the regular season and playoffs, we could see that although LeBron James still did better than average in the semi-circle area under the basket, his shooting significantly declined in that area. It might be the reason why the Lakers in this season did not make it to the finals. Moreover, he still behaved well on the left corner in the playoffs than the other one so the defense strategy against LeBron should be opposite to Stephen Curry.

### Analysis on shooting behavior in different periods of LeBron James 

Every basketball game has four periods and athletes would get exhausted when the game is getting to the end, so it is important to arrange the rest and the rush hour properly. As we previously discovered in Part 5.1.3, the number of 2-point and 3-point shots are similar at the general league level. However, it may differ on the individual player level. Therefore, let us take a look at the shooting behavior of LeBron James in different periods and hope to find some special patterns in it.

```{r lebron_period}
shot_data %>%
  filter(player_name == "LeBron James") %>%
  filter(period!="5" & period!="6") %>%
  mutate(period = factor(period, levels = c("1", "2", "3", "4","Right Side(R)"))) %>%
  ggplot(aes(x = as.numeric(as.character(shot_distance)), group = period, fill = period)) +
  geom_density(alpha = 0.3) +
  theme(legend.position = c(0.9, 0.8)) +
  xlab("Shooting Distance")
```

Based on the plot, LeBron James prefers attacking the rim to shoot long-distance in the second quarter, and among four periods it is the last one that he attempted most threes. Therefore, there is one more defending policy against LeBron that should be focusing more on his layups in the first half and his threes in the second half.

### Shooting Behavior of Giannis Antetokounmpo

Giannis Antetokounmpo, AKA Greek Freak, was experiencing his best seasons in the past two years. MVP winner, All-star, and last but not least, NBA championship. His athleticism is so phenomenal that he could win all even without good shooting ability. While in the shock by shooting ability of Stephen Curry, let us take a look at Giannis Antetokounmpo, another side of basketball.

```{r alluvial_giannis}
shot_data %>%
  filter(player_name == 'Giannis Antetokounmpo') %>%
  mutate(shot_zone_range = factor(shot_zone_range, levels = c("Less Than 8 ft.", "8-16 ft.", "16-24 ft.","24+ ft.","Back Court Shot"))) %>%
  mutate(shot_zone_area = factor(shot_zone_area, levels = c("Left Side Center(LC)", "Right Side Center(RC)", "Center(C)", "Left Side(L)","Right Side(R)","Back Court(BC)"))) %>%
  arrange(event_type) %>%
  ggplot(aes(axis1 = shot_type, axis2 = shot_zone_area, axis3 = shot_zone_range))+
  geom_alluvium(aes(fill = event_type), width = 1/8)+
  geom_stratum(color = 'lightblue', width = 1/8)+
  geom_text(stat = "stratum", aes(label = after_stat(stratum)), size = 2)+
  scale_x_continuous(breaks = 1:3, 
       labels = c("FG Type", "Shot Area", "Shot Range"), expand = c(.01, .05))+
  ggtitle("Alluvial Diagram of Shooting Behavior of Giannis Antetokounmpo")+
  theme(legend.position="bottom",legend.title = element_text(size=5), legend.text = element_text(size=5))+
  ylab('Frequence')
```

Based on this alluvial diagram, we could see that Giannis Antetokounmpo in this season spent most of his attempts on 2 pointers and two-thirds of his shots happened less than 8 feet away from the basket. As a right-handed player, it seems like that he does not have a preference on which side to shoot. Moreover, after comparing the shot attempted at the left side center and right side center, we could find that he approached the rim more often from the left although he is right-handed. This might be inspirational to his defenders that it is better to focus on his left hand. At last, in his fewer long-distance attempts, he did not do very well because we could see more blue flow in the long-distance area. Therefore, he still has a lot of potentials to stimulate and more accomplishments are waiting for him if he could work on his weakness.
