==> 07-conclusion.Rmd <==
# Conclusion

**Concluding the Findings**

From our analysis in chapter 5, we conclude our findings from the shooting pattern for NBA Season 2020-21 as follows:

1. On the general league level, nowadays, basketball tends to shoot fewer mid-range shots - either shooting close to the basket or making more 3-point shots - to try to maximize points per play. Our hypothesis shows that the field goal percentage (whether the goal was made successfully or not) is related to the shooting distance. 

2. Teams vary significantly on their choice for making 2-point or 3-point shots. This preference also differs period by period. As the game approaches the 4th period, teams tend to make more 3-point shots compared to their performance in the 1st period. In addition, there is no clear relation between the FG% and the team's preference for making 2-point or 3-point shots. 

3. Players have different personalities in shooting behavior. 
- Stephen Curry's shooting behavior is a classical example of modern basketball: try less on the mid-range shot and more 2-point shots near the hoop and 3-point shots, and his field goals are mostly above league average. Stephen Curry's weakness is left corner three. 

- LeBron James did fantastic in the paint (near hoop), and he tried a lot of mid-range shots and threes but the results were not ideal. His shooting percentage on the left corner is extremely high, where Stephen Curry was not good at. Comparing Playoffs to Regular Season, LeBron James still did better than average near the hoop, but his shooting significantly declined in that area.

- Giannis Antetokounmpo in this season spent most of his attempts on 2-point shots and 2/3 of his shots happened less than 8 feet away from the basket, and he approached the rim more often from the left although he is right-handed.

**Discussing Limitations**

Although we discussed shooting behaviors on different levels from broad to detail (from league to team and player), we only focused on the shot features - such as shot time, location, distance, etc. - made by each team and player without other information. However, there are still other considerations. For example, in determining the shooting behaviors of a player, the player demographics (height, weight, etc.) and the position he plays (guard or forward) can also affect the player's shooting behavior. Also, other considerations such as the defender's distance from the player can affect the player's shots. In determining the shooting behaviors of a team, it may heavily rely on the kinds of players in that team and the experience and preference of the coach. These are all aspects that we haven't included in our project when determining the shooting behaviors.


**Looking Into the Future**

With the more advanced visualization tools, analyzing the shooting behavior of a team or player becomes easier, and we can explore more underlying insights by using different ways to visualize the numbers. Considering the limitations we discussed before, there is a wide array of aspects to analyze the NBA shooting data and interpreting strategies of a team or a player. With more NBA data being utilized, we look forward that Exploratory Data Analysis on NBA data may in the future help players or teams to see their weaknesses more easily and make improvements, and even find the opponent's weaknesses to improve the defence.
